LayoutTests/http/tests/security/XFrameOptions/x-frame-options-cached-expected.txt

CONSOLE MESSAGE: Invalid 'X-Frame-Options' header encountered when loading 'http://127.0.0.1:8000/security/XFrameOptions/resources/nph-cached-xfo.pl': 'allowall' is not a recognized directive. The header will be ignored.
ALERT: This must fire twice
CONSOLE MESSAGE: Invalid 'X-Frame-Options' header encountered when loading 'http://127.0.0.1:8000/security/XFrameOptions/resources/nph-cached-xfo.pl': 'allowall' is not a recognized directive. The header will be ignored.
ALERT: This must fire twice
Check that an X-Frame-Options header added by a 304 response does not override one from the original request.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Two console messages should be generated, each followed by an alert.
Both frames should point to the expected URL, and not to about:blank.
PASS frame1.src is "http://127.0.0.1:8000/security/XFrameOptions/resources/nph-cached-xfo.pl"
PASS frame2.src is "http://127.0.0.1:8000/security/XFrameOptions/resources/nph-cached-xfo.pl"
PASS successfullyParsed is true

TEST COMPLETE
  
